move groupCardIdWithMetdata and compareTagNames out of views/utils to be shared with store/
