+++ tensorboard/webapp/metrics/views/main_view/card_groups_container.ts
@@ -21,7 +21,7 @@ import {State} from '../../../app_state';
 import {getMetricsFilteredPluginTypes} from '../../store';
 import {CardObserver} from '../card_renderer/card_lazy_loader';
 import {CardGroup} from '../metrics_view_types';
-import {groupCardIdWithMetdata} from '../utils';
+import {groupCardIdWithMetdata} from '../../utils';
 import {getSortedRenderableCardIdsWithMetadata} from './common_selectors';
 
 @Component({
